% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3.R
\name{correlation}
\alias{correlation}
\title{function3}
\usage{
correlation(x)
}
\arguments{
\item{x}{a random year which is numeric}
}
\description{
This function allows you to return a correlation matrix of numeric value of a year.
}
\examples{
function3()
}
\keyword{cats}
