% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_1.R
\name{highest_FTr}
\alias{highest_FTr}
\title{An NBA Function}
\usage{
highest_FTr(year)
}
\arguments{
\item{year}{The year.}
}
\description{
This function allows you to find the best performaned player.
}
\examples{
highest_FTr()
}
\keyword{NBA_FTR}
