% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{mostpoints}
\alias{mostpoints}
\title{function1}
\usage{
mostpoints(x)
}
\arguments{
\item{x}{a random year which is numeric.}
}
\description{
This function allows you to output the one who got most points during a year.
}
\examples{
function1()
}
\keyword{POINTS}
