% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{oldest}
\alias{oldest}
\title{function2}
\usage{
oldest(x)
}
\arguments{
\item{x}{a random year which is numeric}
}
\description{
This function allows you to find out the one who is the oldest furing a year.
}
\examples{
function2()
}
\keyword{AGE}
